+++ Source/iti-react/Changelog.md
@@ -201,3 +201,4 @@ Migration path:
 *   Fix withValidation equality comparison bug, which caused onValidChange to be called infinitely
 *   Fix double-clicking a link cancelling navigation
 *   \*\*\* Remove second type argument from IDataUpdater<>
+*   ThemeColors now includes all theme colors rather than just a few